refact: 해당룩북 불러오기 수정
