+++ src/api/hook/BookHook.ts
@@ -1,4 +1,5 @@
 import { useQuery } from 'react-query';
+import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 
 import { userIdState } from '../../utils/Recoil';
@@ -13,5 +14,16 @@ export const useFindAllBooks = () => {
     queryKey: ['books'],
     enabled: !!userId,
     onError: () => alert('룩북 정보를 불러오는데 실패했습니다.'),
+  });
+};
+
+export const useFindSomeBooks = () => {
+  const { id } = useParams();
+
+  return useQuery({
+    queryFn: () => BookService.getSomeBooks({ id }),
+    queryKey: ['books'],
+    enabled: !!id,
+    onError: () => alert('상세 룩북 정보를 불러오는데 실패했습니다.'),
   });
 };
